code fixes

+++ src/data_access/ParkingLotDAO.java
@@ -43,7 +43,7 @@ public class ParkingLotDAO implements GreenPDAO {
 
                 String id = parseId(parkingLot);
                 String website = parseWebsite(parkingLot);
-                float[] latLong = parseLatLong(parkingLot);
+                double[] latLong = parseLatLong(parkingLot);
                 int capacity = parseCapacity(parkingLot);
                 String streetAddress = parseStreetAddress(parkingLot);
                 String carparkType = parseCarparkType(parkingLot);
@@ -74,10 +74,10 @@ public class ParkingLotDAO implements GreenPDAO {
         return parkingLot.get("slug").toString();
     }
 
-    private float[] parseLatLong(JSONObject parkingLot) {
-        return new float[]{
-                Float.parseFloat(parkingLot.get("lat").toString()),
-                Float.parseFloat(parkingLot.get("lng").toString())
+    private double[] parseLatLong(JSONObject parkingLot) {
+        return new double[]{
+                Double.parseDouble(parkingLot.get("lat").toString()),
+                Double.parseDouble(parkingLot.get("lng").toString())
         };
     }
